MouseSystemEx 1.7.0 -> 1.8.0
* Mouse click, hover and leave activate now activate all events under
the mouse, not just the topmost. This also affects cursor changes and
icons; the topmost event will overwrite those below if conflicts arise.
* Internal "namespace" structure changed for brevity, remains the same
functionally and externally
* Externalized getting events from xy coords to its own method to make
compatibility easier for other plugins that interact with this one
* Code comments added to more complex methods, also as part of the
effort to provide full JSDoc compliant comments for all plugins. This
will be rolled out gradually as it takes time.